pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Bruno Luiz de Deus Adão <brunoblda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.3"
kivy = {extras = ["full"], version = "^2.3.0"}
kivy-examples = "^2.3.0"
pdfservices-sdk = "^4.0.0"
tabula-py = "^2.10.0"
python-dotenv = "^1.0.1"
jpype1 = "^1.5.0"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11.0"
taskipy = "^1.14.0"
ignr = "^2.2"

[tool.taskipy.tasks]
build = "pyinstaller --onefile --paths ./venv/Lib/site-packages main.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
